DREAMWEB: Move sound related code to a new source file
